moved tokenizer and updated train.py
