+++ README.md
@@ -1,1 +1,2 @@
 Building a character level language model just to see how it works
+Intentionally avoiding certain abstractions so I can better understand the architecture
